requirements.txt
Django>=4.2.0,<5.0
django-environ>=0.11.0
djangorestframework>=3.14.0
django-allauth==0.57.0
Pillow>=10.0.0
whitenoise>=6.5.0
redis>=4.6.0
celery>=5.3.0
python-telegram-bot>=20.0
